nbody.cu: memset the g_host{SOA,AOS}_Force arrays before running the test
Without this, it's possible for these arrays to be filled with data from the
crosscheck run, and be untouched by any of the tested functions (e.g. if one of
the tested functions returns 0.0f due to failure, we wouldn't be able to tell
because the arrays contain a perfect match for the crosscheck).

Signed-off-by: Steven Noonan <steven@uplinklabs.net>